Cleaned up code and add functionality for 1 page only orders
